requirements.txt
bleach==3.1.1
Django==2.2
django-debug-toolbar==2.2
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
webencodings==0.5.1
